+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.2.0-rc.1
+
+* Rewrite to support adding cookie to redirected URL.
+* Add `debugLog` parameter to allow printing the debug log.
+* Add `maxRedirects` parameter to set the max redirects.
+
 ## 0.1.0
 
 * Update dependencies